requirements.txt
Flask==3.0.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
scikit-learn==1.3.2
joblib==1.3.2
lime==0.2.0.1
numpy==1.24.3
colorama==0.4.6
